Deleted 48h6e2.edm to solve the old IC issue
The 48h6e2.edm file was still in the shape folder, probably because of a mishap at some point during the merges of various versions.
As with 1.3.0, it caused IC failures for some users. As it is no longer needed because 48h6e2_hds.edm is the file used instead in the 48n6.lua and 48n6e2.lua weapon files, I simply deleted it.