Changed nr. of trackers for 92H6E, "SA-20" replaced by "SA-20A" for all PMU1 units, added tags to all lanchers, edited countries for PS, PMU1 and PMU2 (#30)
* Change from 6 to 36 trackers for the 92H6E

"The S-300PMU-2 is referred in the West as SA-20B Gargoyle. The fire control radar has a range of 300 km, can detect 100 targets, track and engage 36 of them and guide 72 missiles. It is worth noting that earlier radars of the S-300 family could simultaneously engage only 6 targets and guide 12 missiles."

Co-authored-by: Auranis <edward.m.crouch@gmail.com>