Edited D_max range.
There is an issue with our current D_max implementation: we use a pseudo-linear conversion from max range to calculate it (eg. D_max = .3*Range_max). The problem with this issue is that the longer the range (or the max altitude), the bigger the delta between max range and max range at low altitude. This is a problem because this isn't the case in real life, where missiles like the 48N6E2 can egage ISR platforms close to their max kinematic range. With the current implementation, an E-3A AWACS flying at FL300 will be engaged at approx:
- 90km for the 48N6 (vs. 150km max kin. range)
- 110km for the 48N6E2 (vs 200km max kin. range)
- etc.

Using a fixed delta between D_max and Range_max can solve this issue. 50km seems to yield realistic results, according to open-source altitude-range diagrams for the S-300PMU1/2/400.